⚡ 리뷰 점수 수정&예외처리 수정시 HTTP 상태 수정

+++ src/main/java/org/example/tentrilliondollars/review/dto/ReviewResponse.java
@@ -12,7 +12,7 @@ public class ReviewResponse {
     private Long id;
     private String content;
     private String photo;
-    private Long score;
+    private int score;
     public ReviewResponse(Review review){
         this.id = review.getId();
         this.content = review.getContent();
